Some progress towards separating header/implementation in loop-analysis/.
