+++ src/loop-analysis/operation-type.hpp
