+++ src/loop-analysis/sparse-analysis.hpp
